とりあえず軽くメニューが出るように

+++ app/src/main/java/com/servebbs/amazarashi/kangtangdotterzero/domains/primitive/DotFont.java
@@ -6,7 +6,7 @@ import com.servebbs.amazarashi.kangtangdotterzero.domains.ScreenSize;
 
 public class DotFont {
 
-    public static final float normalHeight = (float) (ScreenSize.getIconSize() / 2);
+    public static final int normalHeight = ScreenSize.getIconSize() / 2;
     public static final int numericWidth = ScreenSize.getDotSize() * 5;
 
     private static Typeface dotTypeface;
